Chore: Refactor
